build: upgrade typst to 0.11.0-rc1 (#493)
* fix(upstream): upgrade typst to 0.11.0 and fix errors

* dev: sync corpus in typst v0.11.0

* fix(upstream): upgrade typst to 0.11.0 and fix errors

* fix: bump typst version to v0.11.0-rc1

* fix: make clippy happy

* fix: changed hash

* build: update cargo.lock

* build: update cargo.lock

* fix: document links

* fix: unused import

* fix: unused import again

+++ core/src/vector/convert.rs
@@ -171,7 +171,7 @@ impl FromTypst<typst::visualize::Image> for Image {
         Image {
             data: image.data().to_vec(),
             format: format.into(),
-            size: Axes::new(image.width(), image.height()),
+            size: Axes::new(image.width() as u32, image.height() as u32),
             alt: image.alt().map(|s| s.into()),
             hash: Fingerprint::from_u128(hash),
         }
